backend/requirements.txt
psycopg2-binary
Flask
uvicorn
flask-cors